+++ ubiconfig/__init__.py
@@ -2,4 +2,4 @@ from ubiconfig.ubi import get_loader
 from ubiconfig._impl.loaders.base import Loader
 from ubiconfig.config_types import UbiConfig
 
-__all__ = ['get_loader', 'Loader', 'UbiConfig']
+__all__ = ["get_loader", "Loader", "UbiConfig"]