+++ ubiconfig/_impl/loaders/__init__.py
@@ -1,4 +1,4 @@
 from .gitlab import GitlabLoader as _GitlabLoader
 from .local import LocalLoader as _LocalLoader
 
-__all__ = ['_GitlabLoader', '_LocalLoader']
+__all__ = ["_GitlabLoader", "_LocalLoader"]