+++ ubiconfig/config_types/__init__.py
@@ -15,7 +15,8 @@ class UbiConfig(object):
     ContentSets:
       config.content_sets.rpm.input
       config.content_sets.debuginfo.output"""
-    def __init__(self, cs, pkgs, mds, file_name):
+
+    def __init__(self, cs, pkgs, mds, file_name, version):
         """
         :param cs: :class:`~ubiconfig.config_types.content_sets.ContentSetsMapping`
         :param pkgs: :class:`~ubiconfig.config_types.packages.Packages`
@@ -26,12 +27,13 @@ class UbiConfig(object):
         self.packages = pkgs
         self.modules = mds
         self.file_name = file_name
+        self.version = version
 
     def __repr__(self):
         return self.file_name
 
     @classmethod
-    def load_from_dict(cls, data, file_name):
+    def load_from_dict(cls, data, file_name, version=None):
         """Create new instance of UbiConfig and load it from provided dictonary with
         following format:
 
@@ -50,13 +52,15 @@ class UbiConfig(object):
         See also :meth:`ubiconfig.config_types.content_sets.ContentSetsMapping.load_from_dict`
         :meth:`ubiconfig.config_types.modules.Modules.load_from_dict`
         """
-        m_data = Modules.load_from_dict(data.get('modules', {}))
-        pkgs = data.get('packages', {})
-        pkgs_data = Packages(pkgs.get('include', []),
-                             pkgs.get('exclude', []),
-                             data.get('arches', []))
-        cs_map = ContentSetsMapping.load_from_dict(data['content_sets'])
+        m_data = Modules.load_from_dict(data.get("modules", {}))
+        pkgs = data.get("packages", {})
+        pkgs_data = Packages(
+            pkgs.get("include", []), pkgs.get("exclude", []), data.get("arches", [])
+        )
+        cs_map = ContentSetsMapping.load_from_dict(data["content_sets"])
         # use the simplified file name
-        file_name = file_name.split('/')[-1]
+        file_name = file_name.split("/")[-1]
 
-        return cls(cs=cs_map, pkgs=pkgs_data, mds=m_data, file_name=file_name)
+        return cls(
+            cs=cs_map, pkgs=pkgs_data, mds=m_data, file_name=file_name, version=version
+        )