+++ ubiconfig/config_types/packages.py
@@ -4,11 +4,11 @@
 class Package(object):
     def __init__(self, package, arches):
         self.package = package
-        if '.' in package:
-            name, arch = package.rsplit('.', 1)
-            if arch in arches or arch == '*':
+        if "." in package:
+            name, arch = package.rsplit(".", 1)
+            if arch in arches or arch == "*":
                 self.name, self.arch = name, arch
-        if not hasattr(self, 'name'):
+        if not hasattr(self, "name"):
             self.name = package
             self.arch = None
 
@@ -17,10 +17,9 @@ class Package(object):
 
 
 class IncludePackage(Package):
-
     def __init__(self, package, arches):
         super(IncludePackage, self).__init__(package, arches)
-        if '*' in self.name:
+        if "*" in self.name:
             raise ValueError("<name>*.<arch> is not supported in whitelist")
 
 
@@ -29,7 +28,6 @@ class ExcludePackage(Package):
 
 
 class Packages(object):
-
     def __init__(self, include, exclude, arches):
         """
         Args: